pyproject.toml
[project]
name = "dalcedo-tui"
version = "0.1.0"
description = "BigQuery Natural Language Query Interface"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "textual>=0.89.0",
    "anthropic>=0.40.0",
    "openai>=1.50.0",
    "google-genai>=1.0.0",
    "duckdb>=1.1.0",
    "google-cloud-bigquery>=3.25.0",
    "google-cloud-bigquery-storage>=2.26.0",
    "pyarrow>=17.0.0",
    "xdg-base-dirs>=6.0.0",
    "tomli-w>=1.0.0",
    "rich>=13.9.0",
    "pyyaml>=6.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=5.0.0",
    "textual-dev>=1.5.0",
    "ruff>=0.6.0",
    "mypy>=1.11.0",
]

[project.scripts]
dalcedo = "dalcedo.__main__:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/dalcedo"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.mypy]
python_version = "3.11"
strict = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
